PR리뷰 반영 - 나눗셈 할 때 toDouble은 한군데만 https://github.com/next-step/kotlin-lotto/pull/793#discussion_r1252156079

+++ src/main/kotlin/lotto/domain/WinningStatistics.kt
@@ -11,7 +11,7 @@ class WinningStatistics(val winningStatistics: Map<Rank, Int>) {
 
     fun calculateProfitRate(purchaseMoney: Long): Double {
         val totalPrizeMoney = calculateTotalPrizeMoney()
-        return totalPrizeMoney.toDouble() / purchaseMoney.toDouble()
+        return totalPrizeMoney / purchaseMoney.toDouble()
     }
 
     companion object {
